changed how expand/collapse on object_reference is presented (once expand/collapse is re-enabled for related/associated objects, this will apply there too)
instead of being a black bar saying "toggle" the heading is now linked to the toggle action and there is a "+" sign placed at the front of the line via css.

the highlight and jumpt to target has been moved to a separate link at the end of the line.